add analyze and server_benchmark feature
--analyze uses malt to trace the benchmarks behavior. It uses the run
loop but the obtained results are not stored

Benchmark.server_benchmark is used if only a server is started for
each allocator and clients are used to measure its performance in the run loop.
If server_benchmark is set to True the cmds are run with the system default
allocator.

Misc changes:

* The global environment is no longer changed. Instead a custom env dict
is passed to suprocesses containing LD_PRELOAD.

* Failing cmds no longer skip the whole benchmark instead they now skip the
malfunctioning allocator.

* Fix default title in plot_single_arg

an analyse run are not stored