introduce server concept to Benchmark
A benchmark object can specify a list of cmds to execute as
"servers" in the member attribute server_cmds.
Servers are started and terminated through Popen objects.
This requires the server cmds to not daemonize so the server
can be terminated through the Popen object.
For each started server cmd a shutdown function is registered with
atexit to terminate all servers even if a exception occurs.

Use the new server concept in httpd and mysql benchmarks