fix benchmarks for any glibc version
use the included loader of the build glibc instead of LD_LIBRARY_PATH.
Using LD_LIBRARY_PATH probably fails because of incompatible system
loaders and newer glibc versions.
Then we have to reverse the positions of {measure cmd} and {cmd prefix}
in the exec chain because ld loads only ELF binaries and {measure cmd}
could be a script.