increase workload size for bigger max speedup
with the current workload size of 100K allocations the parallel part (worker)
takes on my development laptop 96% of the single threaded execution time
resulting in a maximal speedup of 25. (Amdahl: 1/(1-0.96) ~= 25)
Increasing the the workload to 1M allocations ups the parallel portion
to ~99% which gives us a max speedup of ~295 enough for all our test machines.
With the new workload the single threaded execution time is ~100s on my
laptop and ~30s on one of the test servers which is tolerable in my
opinion.